requirements.txt
fastapi
uvicorn
python-multipart
pydantic
aiofiles
markdown
python-dotenv
google-genai
chromadb
watchdog
keyring
httpx
beautifulsoup4
lxml
markdownify
feedparser
neo4j
psutil
pymupdf
youtube-transcript-api==1.2.4
google-auth==2.50.0
google-auth-httplib2==0.3.1
brotli==1.2.0
